add isolated modules flag in tsconfig, create consts files for enums, fix import errors
