+++ src/app/providers/StoreProvider/index.ts
@@ -2,15 +2,11 @@ import { createReduxStore, AppDispatch } from './config/store';
 import { StoreProvider } from './ui/StoreProvider';
 import type {
     StateSchema,
-    ReduxStoreWithManager,
     ThunkConfig,
+    ReduxStoreWithManager,
 } from './config/StateSchema';
 
-export {
-    StoreProvider,
-    createReduxStore,
-    StateSchema,
-    ReduxStoreWithManager,
-    AppDispatch,
-    ThunkConfig,
-};
+export { StoreProvider, createReduxStore, ReduxStoreWithManager };
+
+export type { StateSchema, AppDispatch, ThunkConfig };
+
