+++ src/entities/Article/model/consts/articleConsts.ts
@@ -1,0 +1,23 @@
+export enum ArticleSortField {
+    VIEWS = 'views',
+    TITLE = 'title',
+    CREATED = 'created',
+}
+export enum ArticleBlockType {
+    CODE = 'CODE',
+    IMAGE = 'IMAGE',
+    TEXT = 'TEXT',
+}
+
+export enum ArticleType {
+    ALL = 'ALL',
+    IT = 'IT',
+    SCIENCE = 'SCIENCE',
+    ECONOMICS = 'ECONOMICS',
+}
+
+
+export enum ArticleView {
+    LIST = 'LIST',
+    GRID = 'GRID',
+}
