+++ src/entities/Article/ui/ArticleSortSelector/ArticleSortSelector.tsx
@@ -1,4 +1,4 @@
-import { ArticleSortField } from '../../model/types/article';
+import { ArticleSortField } from '../../model/consts/articleConsts';
 import { FC, memo, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { classNames } from 'shared/lib/classnames/classNames';
