+++ src/entities/Article/ui/ArticleTypeTabs/ArticleTypeTabs.tsx
@@ -1,4 +1,4 @@
-import { ArticleType } from '../../model/types/article';
+import { ArticleType } from '../../model/consts/articleConsts';
 import { FC, memo, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { classNames } from 'shared/lib/classnames/classNames';
