+++ src/entities/Counter/index.ts
@@ -1,4 +1,5 @@
 import { counterReducer } from './model/slice/counterSlice';
 import { CounterSchema } from './model/types/counterSchema';
 import { Counter } from './ui/Counter';
-export { Counter, counterReducer, CounterSchema };
+export { Counter, counterReducer };
+export type { CounterSchema };
