+++ src/entities/Profile/index.ts
@@ -1,3 +1,3 @@
 export { ProfileCard } from './ui/ProfileCard/ProfileCard';
 
-export { Profile } from './model/types/profile';
+export type { Profile } from './model/types/profile';
