+++ src/entities/User/index.ts
@@ -1,6 +1,6 @@
 export { userActions, userReducer } from './model/slice/userSlice';
 
-export { UserSchema, User, UserRole } from './model/types/user';
+export type { UserSchema, User } from './model/types/user';
 
 export { getUserAuthData } from './model/selectors/getUserAuthData/getUserAuthData';
 
@@ -11,3 +11,5 @@ export {
     isUserManager,
     getUserRoles,
 } from './model/selectors/roleSelectors';
+
+export { UserRole } from './model/consts/userConsts';
