+++ src/entities/User/model/consts/userConsts.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+    USER = 'USER',
+    ADMIN = 'ADMIN',
+    MANAGER = 'MANAGER',
+}
