+++ src/entities/User/model/selectors/roleSelectors.ts
@@ -1,9 +1,8 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { StateSchema } from 'app/providers/StoreProvider';
-import { UserRole } from '../types/user';
+import { UserRole } from '../consts/userConsts';
 
 export const getUserRoles = (state: StateSchema) => state.user.authData?.role;
-
 
 export const isUserAdmin = createSelector(getUserRoles, (roles) =>
     Boolean(roles?.includes(UserRole.ADMIN))
