+++ src/features/AddCommentForm/index.ts
@@ -1,3 +1,3 @@
-export { AddCommentFormSchema } from './model/types/addCommentform';
+export type { AddCommentFormSchema } from './model/types/addCommentform';
 export { AddCommentFormFormAsync as AddCommentForm } from './ui/AddCommentForm/AddCommentForm.async';
 export { addCommentFormReducer } from './model/slices/addCommentFormSlice';
