+++ src/features/AuthByUsername/index.ts
@@ -1,4 +1,5 @@
 import { LoginModal } from './ui/LoginModal/LoginModal';
 import { LoginSchema } from './model/types/loginSchema';
 
-export { LoginModal, LoginSchema };
+export { LoginModal };
+export type { LoginSchema };
