+++ src/features/EditableProfileCard/index.ts
@@ -1,5 +1,3 @@
 export { EditableProfileCard } from './ui/EditableProfileCard/EditableProfileCard';
-export {
-    ProfileSchema,
-    ValidationProfileError,
-} from './model/types/EditableProfileCardSchema';
+export type { ProfileSchema } from './model/types/EditableProfileCardSchema';
+export { ValidationProfileError } from './model/consts/editableProfileCardsConsts';
