+++ src/features/EditableProfileCard/model/types/EditableProfileCardSchema.ts
@@ -1,14 +1,6 @@
-
 import { Profile } from '/entities/Profile';
+import { ValidationProfileError } from '../consts/editableProfileCardsConsts';
 
-
-export enum ValidationProfileError {
-    INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
-    INCORRECT_AGE = 'INCORRECT_AGE',
-    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
-    NO_DATA = 'NO_DATA',
-    SERVER_ERROR = 'SERVER_ERROR',
-}
 export interface ProfileSchema {
     data?: Profile;
     form?: Profile;
