+++ src/features/UI/index.ts
@@ -2,4 +2,4 @@ export { uiReducer, uiActions } from 'features/UI/model/slices/UISlice';
 
 export { getUIScrollPositionByPath } from './model/selectors/ui';
 
-export { UISchema } from './model/types/UISchema';
+export type { UISchema } from './model/types/UISchema';
