+++ src/pages/ArticleDetailsPage/index.ts
@@ -1,5 +1,5 @@
-export { ArticleDetailsPageSchema } from './model/types';
-export { ArticleDetailsRecommendationsSchema } from './model/types/ArticleDetailsRecommendationsSchema';
+export type { ArticleDetailsPageSchema } from './model/types';
+export type { ArticleDetailsRecommendationsSchema } from './model/types/ArticleDetailsRecommendationsSchema';
 export { ArticleDetailsPageAsync as ArticleDeatilsPage } from './ui/ArticleDetailsPage/ArticleDetailesPage.async';
 
-export { ArticleDetailsCommentsSchema } from './model/types/ArticleDetailsCommentsSchema';
+export type { ArticleDetailsCommentsSchema } from './model/types/ArticleDetailsCommentsSchema';
