+++ src/pages/ArticlesPage/index.ts
@@ -1,1 +1,3 @@
 export { ArticlesPageAsync as ArticlesPage } from './ui/ArticlesPage/ArticlesPage.async';
+
+export type { ArticlePageSchema } from './model/types/articlePageSchema';
