+++ src/pages/ArticlesPage/ui/ArticlesPageFilters/ArticlesPageFilters.tsx
@@ -23,7 +23,7 @@ import cls from './ArticlesPageFilters.module.scss';
 import { SortOrder } from 'shared/types';
 import { fetchArticlesList } from '../../model/services/fetchArticlesList/fetchArticlesList';
 import { useDebounce } from 'shared/lib/hooks/useDebounce/useDebounce';
-import { ArticleType } from 'entities/Article/model/types/article';
+import { ArticleType } from 'entities/Article';
 
 import { ArticleTypeTabs } from 'entities/Article/ui/ArticleTypeTabs/ArticleTypeTabs';
