+++ src/pages/ForbiddenPage/index.ts
@@ -1,0 +1,1 @@
+export {} from './ui/ForbiddenPage';
