+++ src/pages/ProfilePage/index.ts
@@ -1,4 +1,4 @@
-export { ArticlePageSchema } from './../ArticlesPage/model/types/articlePageSchema';
+export type { ArticlePageSchema } from './../ArticlesPage/model/types/articlePageSchema';
 export { ProfilePageAsync as ProfilePage } from './ui/ProfilePage.async';
